ruff format

+++ src/filterbank.py
@@ -1,17 +1,18 @@
 from external.filterbank import filterbank
 import torch
 
+
 def get_filterbank(N: int, sigma: float, device: str) -> torch.Tensor:
     filters, frequency_domain_sum = filterbank(N, sigma)
 
-    filters = filters.astype('complex64')
-    frequency_domain_sum = frequency_domain_sum.astype('float32')
+    filters = filters.astype("complex64")
+    frequency_domain_sum = frequency_domain_sum.astype("float32")
 
     frequency_domain_sum = torch.tensor(frequency_domain_sum, device=device)
     filters = torch.tensor(filters.real, device=device)
 
     filters = filters.view(85, -1)
-    
+
     filters_norm = torch.linalg.vector_norm(filters, dim=-1)
     filters_normalized = filters / filters_norm.view(-1, 1)
     return filters_normalized, frequency_domain_sum
